api/requirements.txt
fastapi == 0.128.0
uvicorn == 0.40.0
pillow == 11.1.0
torch == 2.11.0+cpu
torchvision == 0.26.0+cpu
mlflow == 2.22.0
pydantic == 2.10.6
numpy == 2.1.3
pandas == 2.2.3
python-multipart == 0.0.20
